* changes:
  fix(plat): clang-tidy unused parameter warning
  fix(plat): clang-tidy redundant cast warning
  fix(runtime): clang-tidy redundant cast warnings
  fix(runtime): clang-tidy sizeof-expression warning
  fix(runtime): clang-tidy macro parentheses warnings
  fix(drivers): clang-tidy reserved identifier warning
  fix(runtime): clang-tidy unused parameter warnings

+++ runtime/rmi/realm.c
@@ -125,7 +125,7 @@ static unsigned int s2_num_root_rtts(unsigned int ipa_bits, int sl)
 		return 1U;
 	}
 
-	return (unsigned int)(1U << (ipa_bits - sl_ipa_bits));
+	return (1U << (ipa_bits - sl_ipa_bits));
 }
 
 /*
